REFACTOR: prep for later work
