+++ .gitignore
@@ -26,6 +26,7 @@ testbin/*
 *~
 /kuadrant-operator
 tmp
+topology.dot
 
 /catalog/kuadrant-operator-catalog
 /catalog/kuadrant-operator-catalog.Dockerfile
