+++ api/v1beta2/authpolicy_types.go
@@ -6,7 +6,6 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/google/go-cmp/cmp"
 	authorinoapi "github.com/kuadrant/authorino/api/v1beta2"
-	"github.com/kuadrant/policy-machinery/machinery"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -24,8 +23,6 @@ var (
 		Version: GroupVersion.Version,
 		Kind:    "AuthPolicy",
 	}
-	AuthPoliciesResource = GroupVersion.WithResource("authpolicies")
-	AuthPolicyKind       = schema.GroupKind{Group: GroupVersion.Group, Kind: "AuthPolicy"}
 )
 
 const (
@@ -261,7 +258,6 @@ func (s *AuthPolicyStatus) GetConditions() []metav1.Condition {
 
 var _ kuadrant.Policy = &AuthPolicy{}
 var _ kuadrant.Referrer = &AuthPolicy{}
-var _ machinery.Policy = &AuthPolicy{}
 
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
@@ -358,41 +354,6 @@ func (ap *AuthPolicy) BackReferenceAnnotationName() string {
 
 func (ap *AuthPolicy) DirectReferenceAnnotationName() string {
 	return NewAuthPolicyType().DirectReferenceAnnotationName()
-}
-
-func (ap *AuthPolicy) GetTargetRefs() []machinery.PolicyTargetReference {
-	return []machinery.PolicyTargetReference{
-		machinery.LocalPolicyTargetReference{
-			LocalPolicyTargetReference: ap.Spec.TargetRef,
-			PolicyNamespace:            ap.Namespace,
-		},
-	}
-}
-
-func (ap *AuthPolicy) GetMergeStrategy() machinery.MergeStrategy {
-	return func(policy machinery.Policy, _ machinery.Policy) machinery.Policy {
-		return policy
-	}
-}
-
-func (ap *AuthPolicy) Merge(other machinery.Policy) machinery.Policy {
-	return other
-}
-
-func (ap *AuthPolicy) GetURL() string {
-	return machinery.UrlFromObject(ap)
-}
-
-func (ap *AuthPolicySpec) CommonSpec() *AuthPolicyCommonSpec {
-	if ap.Defaults != nil {
-		return ap.Defaults
-	}
-
-	if ap.Overrides != nil {
-		return ap.Overrides
-	}
-
-	return &ap.AuthPolicyCommonSpec
 }
 
 //+kubebuilder:object:root=true
