+++ api/v1beta2/ratelimitpolicy_types.go
@@ -23,7 +23,6 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/kuadrant/policy-machinery/machinery"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -45,8 +44,6 @@ var (
 		Version: GroupVersion.Version,
 		Kind:    "RateLimitPolicy",
 	}
-	RateLimitPoliciesResource = GroupVersion.WithResource("ratelimitpolicies")
-	RateLimitPolicyKind       = schema.GroupKind{Group: GroupVersion.Group, Kind: "RateLimitPolicy"}
 )
 
 // ContextSelector defines one item from the well known attributes
@@ -218,7 +215,6 @@ func (s *RateLimitPolicyStatus) GetConditions() []metav1.Condition {
 var _ kuadrant.Policy = &RateLimitPolicy{}
 var _ kuadrant.Referrer = &RateLimitPolicy{}
 var _ kuadrantgatewayapi.Policy = &RateLimitPolicy{}
-var _ machinery.Policy = &RateLimitPolicy{}
 
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
@@ -303,29 +299,6 @@ func (r *RateLimitPolicy) BackReferenceAnnotationName() string {
 
 func (r *RateLimitPolicy) DirectReferenceAnnotationName() string {
 	return NewRateLimitPolicyType().DirectReferenceAnnotationName()
-}
-
-func (r *RateLimitPolicy) GetTargetRefs() []machinery.PolicyTargetReference {
-	return []machinery.PolicyTargetReference{
-		machinery.LocalPolicyTargetReference{
-			LocalPolicyTargetReference: r.Spec.TargetRef,
-			PolicyNamespace:            r.Namespace,
-		},
-	}
-}
-
-func (r *RateLimitPolicy) GetMergeStrategy() machinery.MergeStrategy {
-	return func(policy machinery.Policy, _ machinery.Policy) machinery.Policy {
-		return policy
-	}
-}
-
-func (r *RateLimitPolicy) Merge(other machinery.Policy) machinery.Policy {
-	return other
-}
-
-func (r *RateLimitPolicy) GetURL() string {
-	return machinery.UrlFromObject(r)
 }
 
 // CommonSpec returns the Default RateLimitPolicyCommonSpec if it is defined.
